+++ include/colors.h
@@ -20,6 +20,7 @@ IN THE SOFTWARE.
 #define _color_h_
 
 #include "functions.h"
+#include <iostream>
 
 using namespace std;
 
@@ -67,6 +68,8 @@ struct Color {
    * An extimate of luminosity, evaluted as mean between the maximum and the minimum value
    */
   float luminosity() { return (fmax(fmax(r, g), b) + fmin(fmin(r, g), b)) / 2.0; }
+  
+  void print(){ cout << "Red: "<< r << ", Green: " << g << ", Blue: " <<b<<endl;}
 };
 
 //––––––––––––––––––––– Color white and black ––––––––––––––––––––––––––––––––––