+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## HEAD
 
+- Add documentation [#20](https://github.com/ElisaLegnani/PhotorealisticRendering/pull/20)
 - Implement new shapes for the demo image [#21](https://github.com/ElisaLegnani/PhotorealisticRendering/pull/21)
+
+### 2021-7-01
+
+- Add tests for some for some (not tested yet) functionalities [#24](https://github.com/ElisaLegnani/PhotorealisticRendering/pull/24)
 
 ### 2021-6-30