+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ## HEAD
 
-- Fix a bug in sphere rendering [#26](https://github.com/ElisaLegnani/PhotorealisticRendering/pull/26)
 - Add documentation [#20](https://github.com/ElisaLegnani/PhotorealisticRendering/pull/20)
 - Implement new shapes for the demo image [#21](https://github.com/ElisaLegnani/PhotorealisticRendering/pull/21)
+
+### 2021-7-08
+
+- Fix a bug in DiffuseBRDF ray scattering [#26](https://github.com/ElisaLegnani/PhotorealisticRendering/pull/26)
 
 ### 2021-7-01